[SPARK-49960] move non join transformer tests to ExpressionEncoderSuite, the types are proved by round tripping anyway
