+++ sql/core/src/test/scala/org/apache/spark/sql/DatasetSuite.scala
@@ -23,7 +23,7 @@ import java.sql.{Date, Timestamp}
 import scala.collection.immutable.HashSet
 import scala.collection.mutable
 import scala.jdk.CollectionConverters._
-import scala.reflect.{classTag, ClassTag}
+import scala.reflect.ClassTag
 import scala.util.Random
 
 import org.apache.hadoop.fs.{Path, PathFilter}
@@ -38,13 +38,11 @@ import org.apache.spark.scheduler.{SparkListener, SparkListenerJobStart}
 import org.apache.spark.sql.catalyst.{FooClassWithEnum, FooEnum, ScroogeLikeExample}
 import org.apache.spark.sql.catalyst.DeserializerBuildHelper.createDeserializerForString
 import org.apache.spark.sql.catalyst.SerializerBuildHelper.createSerializerForString
-import org.apache.spark.sql.catalyst.encoders.{AgnosticEncoder, AgnosticEncoders, AgnosticExpressionPathEncoder, Codec, ExpressionEncoder, OuterScopes}
-import org.apache.spark.sql.catalyst.encoders.AgnosticEncoders.{BoxedIntEncoder, EncoderField, IterableEncoder, MapEncoder, OptionEncoder, PrimitiveIntEncoder, ProductEncoder, TimestampEncoder, TransformingEncoder}
+import org.apache.spark.sql.catalyst.encoders.{AgnosticEncoders, AgnosticExpressionPathEncoder, ExpressionEncoder, OuterScopes}
+import org.apache.spark.sql.catalyst.encoders.AgnosticEncoders.{BoxedIntEncoder, ProductEncoder}
 import org.apache.spark.sql.catalyst.expressions.{CodegenObjectFactoryMode, Expression, GenericRowWithSchema}
 import org.apache.spark.sql.catalyst.plans.JoinType
 import org.apache.spark.sql.catalyst.trees.DataFrameQueryContext
-import org.apache.spark.sql.catalyst.util.DateTimeUtils.microsToInstant
-import org.apache.spark.sql.catalyst.util.SparkDateTimeUtils.instantToMicros
 import org.apache.spark.sql.catalyst.util.sideBySide
 import org.apache.spark.sql.execution.{LogicalRDD, RDDScanExec, SQLExecution}
 import org.apache.spark.sql.execution.adaptive.AdaptiveSparkPlanHelper
@@ -2860,179 +2858,6 @@ class DatasetSuite extends QueryTest
     checkDataset(Seq(seqMutableSet).toDS(), seqMutableSet)
     checkDataset(Seq(mapMutableSet).toDS(), mapMutableSet)
   }
-
-  // below tests are related to SPARK-49960 and TransformingEncoder usage
-  test("""Encoder with OptionEncoder of transformation""".stripMargin) {
-    type T = Option[V[V[Int]]]
-    val data: Seq[T] = Seq(None, Some(V(V(1))))
-
-    /* attempt to behave as if value class semantics except the last product,
-      using a final transforming instead of a product serializes */
-    val enc =
-      OptionEncoder(
-        transforming(
-          V_OF_INT,
-          true
-        )
-      )
-
-    val sparkViaAgnostic = {
-      val ds = spark.createDataset(data)(enc)
-      ds.schema
-    }
-
-    /* The schema has been changed to just the Product V[Int], the wrapping Option value
-        struct has been removed - it should not have been */
-    assert(sparkViaAgnostic === enc.schema)
-
-    val ds = spark.createDataset(data)(enc)
-    assert(ds.collect().toVector === data.toVector)
-  }
-
-  def provider[A]: () => Codec[V[A], A] = () =>
-    new Codec[V[A], A]{
-      override def encode(in: V[A]): A = in.v
-      override def decode(out: A): V[A] = if (out == null) null else V(out)
-    }
-
-  def transforming[A](underlying: AgnosticEncoder[A],
-                      useUnderyling: Boolean = false): TransformingEncoder[V[A], A] =
-    TransformingEncoder[V[A], A](
-      implicitly[ClassTag[V[A]]],
-      underlying,
-      provider,
-      if (useUnderyling) {
-        underlying.nullable
-      } else {
-        false
-      }
-    )
-
-  val V_INT = StructType(Seq(StructField("v", IntegerType, nullable = true)))
-
-  // product encoder for a non-nullable V
-  val V_OF_INT =
-    ProductEncoder(
-      classTag[V[Int]],
-      Seq(EncoderField("v", PrimitiveIntEncoder, nullable = false, Metadata.empty)),
-      None
-    )
-
-  test("""Encoder derivation with nested TransformingEncoder of OptionEncoder""".stripMargin) {
-    type T = V[V[Option[V[Int]]]]
-    val data: Seq[T] = Seq(V(V(None)), V(V(Some(V(1)))))
-
-    /* attempt to behave as if value class semantics except the last product,
-      using a final transforming instead of a product serializes */
-    val enc =
-      transforming(
-        transforming(
-          OptionEncoder(
-            V_OF_INT
-          )
-        )
-      )
-
-    val sparkViaAgnostic = {
-      val ds = spark.createDataset(data)(enc)
-      ds.schema
-    }
-
-    /* The schema has been changed to just the Product V[Int], the wrapping Option value
-        struct has been removed - it should not have been */
-    assert(sparkViaAgnostic === enc.schema)
-
-    val ds = spark.createDataset(data)(enc)
-    assert(ds.collect().toVector === data.toVector)
-  }
-
-  test("""Encoder derivation with TransformingEncoder of OptionEncoder""".stripMargin) {
-    type T = V[Option[V[Int]]]
-    val data: Seq[T] = Seq(V(None), V(Some(V(1))))
-
-    /* attempt to behave as if value class semantics except the last product,
-      using a final transforming instead of a product serializes */
-    val enc =
-      transforming(
-        OptionEncoder(
-          V_OF_INT
-        )
-      )
-
-    val sparkViaAgnostic = {
-      val ds = spark.createDataset(data)(enc)
-      ds.schema
-    }
-
-    /* The schema has been changed to just the Product V[Int], the wrapping Option value
-        struct has been removed - it should not have been */
-    assert(sparkViaAgnostic === enc.schema)
-
-    val ds = spark.createDataset(data)(enc)
-    assert(ds.collect().toVector === data.toVector)
-  }
-
-  val longEncForTimestamp: AgnosticEncoder[V[Long]] =
-    TransformingEncoder[V[Long], java.sql.Timestamp](
-      classTag,
-      TimestampEncoder(true),
-      () =>
-        new Codec[V[Long], java.sql.Timestamp] with Serializable {
-          override def encode(in: V[Long]): Timestamp = Timestamp.from(microsToInstant(in.v))
-
-          override def decode(out: Timestamp): V[Long] = V[Long](instantToMicros(out.toInstant))
-        }
-    )
-
-  test("""TransformingEncoder as Iterable""".stripMargin) {
-    type T = Seq[V[Long]]
-    val data: Seq[T] = Seq(Seq(V(0)), Seq(V(1), V(2)))
-
-    /* requires validateAndSerializeElement to test for TransformingEncoder */
-    val enc: AgnosticEncoder[T] =
-      IterableEncoder[Seq[V[Long]], V[Long]](
-        implicitly[ClassTag[Seq[V[Long]]]],
-        longEncForTimestamp,
-        containsNull = false,
-        lenientSerialization = false)
-
-    assert(enc.dataType === new ArrayType(TimestampType, false))
-
-    val sparkViaAgnostic = {
-      val ds = spark.createDataset(data)(enc)
-      ds.schema
-    }
-
-    assert(sparkViaAgnostic === enc.schema)
-
-    val ds = spark.createDataset(data)(enc)
-    assert(ds.collect().toVector === data.toVector)
-  }
-
-  test("""TransformingEncoder as Map Key/Value""".stripMargin) {
-    type T = Map[V[Long], V[Long]]
-    val data: Seq[T] = Seq(Map(V(0L) -> V(0L)), Map(V(1L) -> V(1L)), Map(V(2L) -> V(2L)))
-
-    /* requires validateAndSerializeElement to test for TransformingEncoder */
-    val enc: AgnosticEncoder[T] =
-      MapEncoder[T, V[Long], V[Long]](
-        implicitly[ClassTag[T]],
-        longEncForTimestamp,
-        longEncForTimestamp,
-        valueContainsNull = false)
-
-    assert(enc.dataType === new MapType(TimestampType, TimestampType, false))
-
-    val sparkViaAgnostic = {
-      val ds = spark.createDataset(data)(enc)
-      ds.schema
-    }
-
-    assert(sparkViaAgnostic === enc.schema)
-
-    val ds = spark.createDataset(data)(enc)
-    assert(ds.collect().toVector === data.toVector)
-  }
 }
 
 /**
@@ -3208,5 +3033,3 @@ case class SaveModeArrayCase(modes: Array[SaveMode])
 
 case class K1(a: Long)
 case class K2(a: Long, b: Long)
-
-case class V[A](v: A)
